show command info privide a unexpected title.20160622-01:13

+++ icurve_v2.0/icurve/source/icv_command.cpp
@@ -171,7 +171,7 @@ qint16 IcvCommand::getState()
 
 QString IcvCommand::getCommandTitle()
 {
-    QString lineIdStr = QString::number(lineId);
+    QString lineIdStr = QString::number((int)lineId);
     QString    title  = name.remove("get") +".line " +  lineIdStr +\
         ((direction == 0)? ".US": ".DS");
